# contracts/openapi/operations.v1.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cutover operations API",
    "version": "1.0.0",
    "description": "Local synthetic operations. Every request checks role and site. Errors use RFC 9457: 400 malformed, 401 unauthenticated, 403 role denied, 404 outside site or absent, 409 conflict, 413 body over 64 KiB, 422 domain validation, 429 rate limit, 503 durable admission unavailable with Retry-After."
  },
  "servers": [
    {
      "url": "http://localhost:8780"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/sites/{siteId}/equipment": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "getEquipment",
        "responses": {
          "200": {
            "description": "Current owner projection, with observation time where applicable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/equipment-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/commands/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getCommand",
        "responses": {
          "200": {
            "description": "Current owner projection, with observation time where applicable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/command-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/commands": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        }
      ],
      "get": {
        "operationId": "listRecoveryCommands",
        "summary": "Read up to 100 site-scoped uncertain or quarantined commands, oldest first",
        "responses": {
          "200": {
            "description": "Observed recovery queue"
          },
          "400": {
            "$ref": "#/components/responses/problem"
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "409": {
            "$ref": "#/components/responses/problem"
          },
          "413": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          },
          "503": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/commands/{id}/reconciliation": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "investigateCommand",
        "summary": "Supervisor records a versioned status investigation using the original command ID",
        "description": "Requires the supervisor role and a reason. A successful response records a request; it is not physical retry authority or proof of completion. A live lease, terminal state or changed version returns 409.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reconciliation-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investigation request recorded; repeats with the same key return the same response"
          },
          "400": {
            "$ref": "#/components/responses/problem"
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "409": {
            "$ref": "#/components/responses/problem"
          },
          "413": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          },
          "503": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/execution-tasks": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "listExecutionTasks",
        "description": "Operator/supervisor read of the independent execution task owner; at most 100 tasks ordered by characterized priority, eligibility and stable movement ID.",
        "responses": {
          "200": {
            "description": "Independent execution task owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/execution-task-list"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/execution-tasks/{id}/recovery": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "recoverExecutionTask",
        "description": "Supervisor resumes an exhausted execution transport investigation with expected version and reason. No allocation or physical command identity is replaced.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/reconciliation-request.v1.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audited status investigation requested"
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/zones": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "listZoneRoutes",
        "description": "Site-scoped operator/supervisor observation. Retained sessions are bounded; observation times identify stale data.",
        "responses": {
          "200": {
            "description": "Current migration observations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone-route-list"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/commands/{id}/timeline": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read retained adapter transitions. An incomplete seven-day event history never invalidates or invents the lifetime command proof.",
        "responses": {
          "200": {
            "description": "Read retained adapter transitions. An incomplete seven-day event history never invalidates or invents the lifetime command proof.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/command-timeline"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/execution-tasks/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read one task from its owning service, scoped to the authenticated site.",
        "responses": {
          "200": {
            "description": "Current owner task projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/execution-task-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "reconciliation-request": {
        "$ref": "../schemas/reconciliation-request.v1.json"
      },
      "problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
      "equipment-view": {
        "$ref": "../schemas/equipment-view.v1.json"
      },
      "command-view": {
        "$ref": "../schemas/command-view.v1.json"
      },
...
      "zone-route-list": {
        "$ref": "../schemas/zone-route-list.v1.json"
      },
...
      "execution-task-list": {
        "$ref": "../schemas/execution-task-list.v1.json"
      },
...
      "command-timeline": {
        "type": "object",
        "required": [
          "commandId",
          "movementId",
          "siteId",
          "observedAt",
          "state",
          "commandVersion",
          "journalRecordedAt",
          "retentionDays",
          "historyComplete",
          "events"
        ],
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "movementId": {
            "type": "string",
            "format": "uuid"
          },
          "siteId": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "state": {
            "type": "string"
          },
          "commandVersion": {
            "type": "integer",
            "minimum": 0
          },
          "journalRecordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "retentionDays": {
            "type": "integer",
            "const": 7
          },
          "historyComplete": {
            "type": "boolean"
          },
          "events": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "type": "object",
              "required": [
                "id",
                "type",
                "version",
                "at",
                "state"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": {
                  "type": "string"
                },
                "version": {
                  "type": "integer",
                  "minimum": 1
                },
                "at": {
                  "type": "string",
                  "format": "date-time"
                },
                "state": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "additionalProperties": true
            }
          }
        },
        "additionalProperties": true
      },
...
      "execution-task-view": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movementId": {
            "type": "string",
            "format": "uuid"
          },
          "allocationId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "epoch": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "eligibleAt": {
            "type": "string",
            "format": "date-time"
          },
          "transportFailures": {
            "type": "integer",
            "minimum": 0
          },
          "transportPaused": {
            "type": "boolean"
          },
          "dispatchAcceptedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "movementId",
          "allocationId",
          "zoneId",
          "state",
          "owner",
          "epoch",
          "version",
          "lastError",
          "eligibleAt",
          "transportFailures",
          "transportPaused",
          "dispatchAcceptedAt"
        ],
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "problem": {
        "description": "Bounded problem response without internal exceptions",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/problem"
            }
          }
        }
      }
    }
  }
}